Fix ECDSASignatureSpec test type mismatch
Changed lines 27 and 42 from BigInt(pointSign) to pointSign.toByte
to use the (ByteString, ByteString, Byte) constructor.

The test was incorrectly mixing ByteString and BigInt types, which
doesn't match any valid ECDSASignature constructor signature:
- (ByteString, ByteString, Byte) ✓
- (BigInt, BigInt, BigInt) ✓
- (ByteString, ByteString, BigInt) ✗

Fixed to use pointSign.toByte for proper type matching.

+++ crypto/src/test/scala/com/chipprbots/ethereum/crypto/ECDSASignatureSpec.scala
@@ -24,7 +24,7 @@ class ECDSASignatureSpec extends AnyFlatSpec with Matchers with ScalaCheckProper
     val pointSign = 28
 
     val sig =
-      ECDSASignature(signatureRandom, signature, BigInt(pointSign))
+      ECDSASignature(signatureRandom, signature, pointSign.toByte)
 
     sig.publicKey(bytesToSign).isEmpty shouldBe false
   }
@@ -39,7 +39,7 @@ class ECDSASignatureSpec extends AnyFlatSpec with Matchers with ScalaCheckProper
     val pointSign = 0x1f
 
     val sig =
-      ECDSASignature(signatureRandom, signature, BigInt(pointSign))
+      ECDSASignature(signatureRandom, signature, pointSign.toByte)
 
     sig.publicKey(bytesToSign).isEmpty shouldBe true
   }
